iterative search

+++ src/main/java/com/bridgelabz/location/model/LatLng.java
@@ -6,12 +6,12 @@ import javax.persistence.Embeddable;
 @Embeddable
 public class LatLng {
 
-	private float lat;
+	private double lat;
 
 	@Column(name = "lng")
-	private float lon;
+	private double lon;
 
-	public LatLng(float lat, float lng) {
+	public LatLng(double lat, double lng) {
 		this.lat = lat;
 		this.lon = lng;
 	}
@@ -20,19 +20,19 @@ public class LatLng {
 
 	}
 
-	public float getLat() {
+	public double getLat() {
 		return lat;
 	}
 
-	public void setLat(float lat) {
+	public void setLat(double lat) {
 		this.lat = lat;
 	}
 
-	public float getLon() {
+	public double getLon() {
 		return lon;
 	}
 
-	public void setLon(float lon) {
+	public void setLon(double lon) {
 		this.lon = lon;
 	}
